requirements.txt
Flask>=3.0
Pillow>=10.0
requests>=2.31
torch>=2.3
torchvision>=0.18
